Fix getting result from metadata envelope.
When the logic was introduced in https://github.com/vrandezo/lambda-cli/commit/9fcd579c05cd86557ee14ac17aa28c01e7223b8e, it was checking whether the second item in the pair was empty, because that one was populated only when there were errors. Since then however execution time has been added in https://github.com/wikimedia/mediawiki-services-function-orchestrator/commit/cae32fd85c7a4af9a7d30ec8c34cb190271636f1, so the second item in the pair is never empty. The proposed change is to check whether the first element is empty instead (although that might be incorrect if the result was really void, however I could not verify that because there seems to be another problem when trying to evaluate void).

BUG: T308771